Renaming `SuggestTypeAlias` intention
