+++ .scalafmt.conf
@@ -1,12 +1,33 @@
-version=2.0.0-RC6
-maxColumn = 120
-align = most
-continuationIndent.defnSite = 2
-assumeStandardLibraryStripMargin = true
-docstrings = JavaDoc
-lineEndings = preserve
-includeCurlyBraceInSelectChains = false
-danglingParentheses = true
-optIn.annotationNewlines = true
+version = "2.3.2"
 
-rewrite.rules = [SortImports, RedundantBraces]
+align = most
+assumeStandardLibraryStripMargin = true
+danglingParentheses = true
+docstrings = JavaDoc
+lineEndings = unix
+maxColumn = 120
+
+continuationIndent {
+  callSite = 2
+  defnSite = 2
+}
+
+newlines {
+  alwaysBeforeTopLevelStatements = true
+  alwaysBeforeMultilineDef = true
+}
+
+spaces {
+  afterKeywordBeforeParen = true
+  afterSymbolicDefs = true
+  inImportCurlyBraces = true
+}
+
+rewrite.rules = [
+  AsciiSortImports,
+  AvoidInfix,
+  PreferCurlyFors,
+  RedundantBraces,
+  RedundantParens,
+  SortModifiers
+]
