+++ src/main/resources/META-INF/plugin.xml
@@ -62,7 +62,8 @@
                          enabledByDefault="true" language="Scala"/>
         <intentionAction>
             <category>ZIO/Suggestions</category>
-            <className>zio.intellij.intentions.actions.SuggestTypeAliasAction</className>
+            <className>zio.intellij.intentions.suggestions.SuggestTypeAlias</className>
+        </intentionAction>
         </intentionAction>
 
         <!-- test runner -->
