+++ src/main/resources/intentionDescriptions/SuggestTypeAlias/after.scala.template
