+++ src/main/resources/intentionDescriptions/SuggestTypeAlias/before.scala.template
