+++ src/main/resources/intentionDescriptions/SuggestTypeAlias/description.html
