+++ src/main/scala/zio/intellij/inspections/package.scala
@@ -6,21 +6,21 @@ import org.jetbrains.plugins.scala.lang.psi.api.expr._
 import org.jetbrains.plugins.scala.lang.psi.api.statements.params.ScParameter
 
 package object inspections {
-  val zio = Array("zio.ZIO")
+  val zioClasses = Array("zio.ZIO")
 
   def invocation(methodName: String) = new Qualified(methodName == _)
 
-  private[inspections] val `.*>`         = invocation("*>").from(zio)
-  private[inspections] val `.as`         = invocation("as").from(zio)
-  private[inspections] val `.map`        = invocation("map").from(zio)
-  private[inspections] val `.mapError`   = invocation("mapError").from(zio)
-  private[inspections] val `.asError`    = invocation("asError").from(zio)
-  private[inspections] val `.catchAll`   = invocation("catchAll").from(zio)
-  private[inspections] val `.foldCause`  = invocation("foldCause").from(zio)
-  private[inspections] val `.foldCauseM` = invocation("foldCauseM").from(zio)
+  private[inspections] val `.*>`         = invocation("*>").from(zioClasses)
+  private[inspections] val `.as`         = invocation("as").from(zioClasses)
+  private[inspections] val `.map`        = invocation("map").from(zioClasses)
+  private[inspections] val `.mapError`   = invocation("mapError").from(zioClasses)
+  private[inspections] val `.asError`    = invocation("asError").from(zioClasses)
+  private[inspections] val `.catchAll`   = invocation("catchAll").from(zioClasses)
+  private[inspections] val `.foldCause`  = invocation("foldCause").from(zioClasses)
+  private[inspections] val `.foldCauseM` = invocation("foldCauseM").from(zioClasses)
 
   def fromZio(r: ScExpression): Boolean =
-    isOfClassFrom(r, zio)
+    isOfClassFrom(r, zioClasses)
 
   class ZIOMemberReference(refName: String) {
     def unapply(expr: ScExpression): Option[ScExpression] = expr match {
