+++ src/main/scala/zio/intellij/intentions/ZTypeAnnotationIntention.scala
@@ -1,18 +1,20 @@
 package zio.intellij.intentions
 
 import com.intellij.openapi.editor.Editor
+import com.intellij.openapi.util.Iconable
+import javax.swing.Icon
 import org.jetbrains.plugins.scala.codeInsight.intention.types._
 import org.jetbrains.plugins.scala.codeInspection.collections.isOfClassFrom
 import org.jetbrains.plugins.scala.lang.psi.api.base.types.ScTypeElement
 import org.jetbrains.plugins.scala.lang.psi.api.statements._
 import org.jetbrains.plugins.scala.lang.psi.types.ScType
-import zio.intellij.inspections.zio
+import zio.intellij.inspections._
 
-abstract class ZTypeAnnotationIntention extends AbstractTypeAnnotationIntention {
+abstract class ZTypeAnnotationIntention extends AbstractTypeAnnotationIntention with Iconable {
   final override def getText: String = getFamilyName
 
   override protected def descriptionStrategy: Strategy = ZStrategy {
-    case (_, declaredType) => isOfClassFrom(declaredType, zio)
+    case (_, declaredType) => isOfClassFrom(declaredType, zioClasses)
     case _                 => false
   }
 
@@ -20,6 +22,8 @@ abstract class ZTypeAnnotationIntention extends AbstractTypeAnnotationIntention 
     case (te, declaredType) if maybeEditor.isDefined => invoke(te, declaredType, maybeEditor.get)
     case _                                           => false
   }
+
+  override def getIcon(flags: Int): Icon = zio.intellij.ZioIcon
 
   protected def invoke(te: ScTypeElement, declaredType: ScType, editor: Editor): Boolean
 }
