+++ src/main/scala/zio/intellij/package.scala
@@ -1,0 +1,8 @@
+package zio
+
+import com.intellij.openapi.util.IconLoader.getIcon
+import javax.swing.Icon
+
+package object intellij {
+  val ZioIcon: Icon = getIcon("/images/zio-full-color-rgb.svg")
+}
